+++ CHANGELOG.md
@@ -32,6 +32,7 @@
     * New: Add `unhook_default_filters()` as a helper method to unhook plugin's default filters hooked to `text_replace()`
     * New: Add tests for setting defaults
     * New: Add text_to_replace example values to verify replacement to 0 and an empty string are valid
+    * New: Add failing tests for replacements affecting shortcode tags and shortcode attributes (though current behavior may be desired)
     * New: Add new `test_does_not_replace_within_markup_attributes()`
     * Change: Rename old `test_does_not_replace_within_markup_attributes()` to `test_does_not_replace_within_markup_attributes_but_does_between_tags()`
     * Change: Update unit test install script and bootstrap to use latest WP unit test repo